Correct misleading option
Rename option '--backup-timeout' because it is misleading. In reality, it is not a timeout for the whole backup operation, it is a timeout for 'no network packets received from the other side' during the backup operation. Thus renamed in order to do not confuse users. Whole operation timeout could be specified by a global '--timeout' option of prlctl.

Signed-off-by: Dmitry Mishin <dim@virtuozzo.com>
Signed-off-by: Igor Sukhih <igor@virtuozzo.com>
Acked-by: Artem Pavlenko <apavlenko@virtuozzo.com>