Polish the utility output in 'exec' cancellation case
When a guest command is cancelled, we do not wait (and don't care)
when it is finished. Thus we have a situation when the job ret code
is zero and there is no parameter with ret code of the executed
command. This patch handles this case and removes wrong error output
on cancellation.

PSBM-45482