Enable "prlctl delete --force" to unregister running invalid VMs. #PSBM-45622
If the configuration of the VM is corrupted while the VM is running, it remains running.
But its state is reported as "invalid".
The dispatcher refuses to delete this VM because it is running.
"prlctl delete" has "--force" flag to stop and delete a running VM in one command,
but it doesn't work for invalid VMs, because prlctl tries to stop only VMs in states
"running" or "paused".

This patch makes "prlctl delete --force" work for invalid VMs as well.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>