prlsrvctl: change vcmmd set policy message from fail to warning #PSBM-143450
If we change the vcmd policy using the prlsrvctl set --vcmmd-policy command,
any non-zero return code is considered a configuration setting error.
In terms of vcmmd VCMMD_ERROR_POLICY_SET_ACTIVE_VES is only a warning.
In the dispatcher, the code VCMMD_ERROR_POLICY_SET_ACTIVE_VES corresponds to
the code PRL_ERR_RUNNING_VM_OR_CT. So we change the behavior of prlsrvctl and
change the message text for PRL_ERR_RUNNING_VM_OR_CT so that the return code is
treated as a warning.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>